Добавлено удаление напоминания, а также мелкие правки

+++ src/main/java/com/example/reminder/repository/ReminderRepository.java
@@ -3,5 +3,5 @@ package com.example.reminder.repository;
 import com.example.reminder.entity.Reminder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReminderRepository extends JpaRepository<Reminder, Integer> {
+public interface ReminderRepository extends JpaRepository<Reminder, Long> {
 }
